+++ src/aind_dynamic_foraging_models/generative_model/__init__.py
@@ -1,4 +1,4 @@
 """Package for generative models of dynamic foraging behavior"""
 
-from .agent_loss_counting import ForagerLossCounting  # noqa: F401
-from .agent_q_learning import ForagerSimpleQ  # noqa: F401
+from .forager_loss_counting import ForagerLossCounting  # noqa: F401
+from .forager_q_learning import ForagerQLearning  # noqa: F401
